lots of changes: x86, i686, 32bit on 64bit, bullseye
lets commit some work of the last weeks - there are still some bugs
but it seems to basically work and its better to have it in the tree
now ... the changes in detail are:
- support for x86_64 = 64bit intel / amd64
- support for i686 = 32bit intel / i386
- support for 32bit userland on top of a 64bit kernel / system
- preparations for the upcoming debian bullseye
- add chromebook_x86_uefi, especially for running with 32bit user-
  land to run well on chromebooks with only 2gb of ram due to the
  lower memory requirements (64bit needs 1/3 to 1/2 more memory)
- a few fixes and improvements here and there

+++ scripts/create-chroot-stage-02.sh
@@ -7,17 +7,21 @@ export LANG=C
 
 systemctl enable ssh
 systemctl disable fstrim.timer
-systemctl disable fwupd.service
-systemctl disable fwupd-refresh.service
+if [ "${1}" != "bullseye" ]; then
+  systemctl disable fwupd.service
+  systemctl disable fwupd-refresh.service
+fi
 
 # in case you want to enable automatic updates, just comment out the next lines
 # TODO: not sure if the first two are still required
 systemctl disable apt-daily
 systemctl disable apt-daily-upgrade
 systemctl disable apt-daily-upgrade.timer
-systemctl disable unattended-upgrades.service
-sed -i 's,Update-Package-Lists "1",Update-Package-Lists "0",g' /etc/apt/apt.conf.d/10periodic
-sed -i 's,Update-Package-Lists "1",Update-Package-Lists "0",g;s,Unattended-Upgrade "1",Unattended-Upgrade "0",g' /etc/apt/apt.conf.d/20auto-upgrades
+if [ "${1}" != "bullseye" ]; then
+  systemctl disable unattended-upgrades.service
+  sed -i 's,Update-Package-Lists "1",Update-Package-Lists "0",g' /etc/apt/apt.conf.d/10periodic
+  sed -i 's,Update-Package-Lists "1",Update-Package-Lists "0",g;s,Unattended-Upgrade "1",Unattended-Upgrade "0",g' /etc/apt/apt.conf.d/20auto-upgrades
+fi
 
 useradd -c ${2} -d /home/${2} -m -p '$6$sEhhlter$njAiCsaYr7lveaAQCmsABlrGbrVip/lcBUlY2M9DUHfaUh0zSLfcJ4mN0BDqH7bg/2BITbp7BK3qPf8zR.3Ad0' -s /bin/bash ${2}
 usermod -a -G sudo ${2}
@@ -30,7 +34,9 @@ sed -i 's,# en_US ISO-8859-1,en_US ISO-8859-1,g;s,# en_US.UTF-8 UTF-8,en_US.UTF-
 locale-gen
 
 # remove snapd and dmidecode (only on ubuntu) as it crashes on some arm devices on boot
-apt-get -yq remove snapd dmidecode
+if [ "${1}" != "bullseye" ]; then
+  apt-get -yq remove snapd dmidecode
+fi
 
 apt-get -yq auto-remove
 apt-get clean
